+++ site/homepage/content/release-notes.md
@@ -4,7 +4,7 @@ title = "What is new & noteworthy in Hono?"
 description = "Information about changes in recent Hono releases. Includes new features, fixes, enhancements and API changes."
 +++
 
-## 1.10.0
+## 1.10.0 (not yet released)
 
 ### Fixes & Enhancements